Zhihu/Lectures/12模型微调下/data_augmentation/requirements.txt
langchain==0.0.336
openai==0.28.1
pydantic==1.10.12
python-dotenv==1.0.1
Requests==2.31.0
tqdm==4.66.1
weaviate-client